feat: navItems 배열로 분리 및 ROUTES를 constants 폴더로 이동
